utils/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.5
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
